requirements.txt
aiogram>=3.4.1
APScheduler>=3.10.4
python-dotenv>=1.0.1
openai>=1.15.0

gspread>=6.1.2
google-auth>=2.30.0
google-api-python-client>=2.136.0

fpdf2>=2.7.9
aiosqlite>=0.20.0
reportlab>=4.0.0
